Refactor- Move token, api calls out of EventDetails component
